Big update

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: zaddi <zaddi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 10:32:31 by zaddi             #+#    #+#             */
-/*   Updated: 2025/11/10 11:57:03 by zaddi            ###   ########.fr       */
+/*   Updated: 2025/11/13 16:02:25 by zaddi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	size_t	i;
+	int	i;
 
 	if (src == dest)
 		return (dest);
 	else if (src > dest)
 	{
 		i = 0;
-		while (i < n)
+		while (i < (int) n)
 		{
 			((char *)dest)[i] = ((char *)src)[i];
 			i++;
@@ -29,7 +29,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 	}
 	else
 	{
-		i = n - 1;
+		i = (int) n - 1;
 		while (i >= 0)
 		{
 			((char *)dest)[i] = ((char *)src)[i];
